fix: change hardcoded minimal json file path
Because web pages are not redeployed every time, but compiled and then remain unchanged. The package is released once a week, but as long as npm run build->run dist generates nsi.min.json, the front-end display can be refreshed (the front-end display does not include any packaged data and is truly pure explorer).